requirements.txt
# 漫合 PDF - 漫画拼接专用工具
# Python 桌面版

PyQt6>=6.5.0
Pillow>=10.0.0
img2pdf>=0.5.0
pyinstaller>=6.0.0
